+++ test/unit/notice_test.rb
@@ -80,6 +80,28 @@ class NoticeTest < Test::Unit::TestCase
     assert_equal hostname, notice.hostname
   end
 
+  context "custom fingerprint" do
+    should "include nil fingerprint when no fingerprint is specified" do
+      notice = build_notice
+      assert_equal nil, notice.fingerprint
+    end
+
+    should "accept fingerprint as string" do
+      notice = build_notice({ :fingerprint => 'foo' })
+      assert_equal '0beec7b5ea3f0fdbc95d0dd47f3c5bc275da8a33', notice.fingerprint
+    end
+
+    should "accept fingerprint responding to #call" do
+      notice = build_notice({ :fingerprint => mock(:call => 'foo') })
+      assert_equal '0beec7b5ea3f0fdbc95d0dd47f3c5bc275da8a33', notice.fingerprint
+    end
+
+    should "accept fingerprint using #to_s" do
+      notice = build_notice({ :fingerprint => mock(:to_s => 'foo') })
+      assert_equal '0beec7b5ea3f0fdbc95d0dd47f3c5bc275da8a33', notice.fingerprint
+    end
+  end
+
   context "with a backtrace" do
     setup do
       @source = <<-RUBY