+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
   *Mike Mondragon @monde*
 
+* Added a custom grouping option
+
+  *Joshua Wood*
+
 * Added option to run capistrano deploy notification asynchronously
 
  *Sergey Efremov*